fixup! glsl: support compilation of geometry shaders via EXT/ARB_geometry_shader4
Revert max_array_access to unsigned.

"glsl: support compilation of geometry shaders via EXT/ARB_geometry_shader4"
changes max_array_access from unsigned to signed type.

This fixup reverts that change.

I can only assume that this was done because max_array_acces was at some point
compared to a signed int that could be negative, but if there ever was such a
comparison it's not in this patch anymore.

+++ src/glsl/ast_to_hir.cpp
@@ -752,7 +752,7 @@ do_assignment(exec_list *instructions, struct _mesa_glsl_parse_state *state,
 
 	 assert(var != NULL);
 
-	 if (var->max_array_access >= rhs->type->array_size()) {
+	 if (var->max_array_access >= unsigned(rhs->type->array_size())) {
 	    /* FINISHME: This should actually log the location of the RHS. */
 	    _mesa_glsl_error(& lhs_loc, state, "array size must be > %u due to "
 			     "previous access",
@@ -2207,7 +2207,7 @@ get_variable_being_redeclared(ir_variable *var, ast_declaration *decl,
        * FINISHME: required or not.
        */
 
-      const int size = var->type->array_size();
+      const unsigned size = unsigned(var->type->array_size());
       check_builtin_array_max_size(var->name, size, loc, state);
       if ((size > 0) && (size <= earlier->max_array_access)) {
 	 _mesa_glsl_error(& loc, state, "array size must be > %u due to "
@@ -2230,7 +2230,7 @@ get_variable_being_redeclared(ir_variable *var, ast_declaration *decl,
        * FINISHME: required or not.
        */
 
-      const int size = var->type->array_size();
+      const unsigned size = unsigned(var->type->array_size());
       check_builtin_array_max_size(var->name, size, loc, state);
       if ((size > 0) && (size <= earlier->max_array_access)) {
 	 _mesa_glsl_error(& loc, state, "array size must be > %u due to "
